LayoutTests/media/modern-media-controls/status-label/status-label-text-align-expected.txt

Testing the StatusLabel text-align is correctly set when hosted in media controls.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(mediaControls.statusLabel.element).textAlign is "left"

PASS successfullyParsed is true

TEST COMPLETE
